scripts/requirements.txt
aenum==3.1.16
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alabaster==0.7.16
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
bleach==6.2.0
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
certifi @ file:///C:/b/abs_2bz30lzohz/croot/certifi_1754570654565/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
coverage==7.10.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
docstring_parser==0.17.0
docutils==0.19
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
fiona==1.10.1
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
h5py==3.14.0
huggingface-hub==0.34.3
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
joblib==1.5.1
jsonargparse==4.40.1
jsonnet==0.21.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
laspy==2.6.1
lazy_loader==0.4
lightly==1.5.22
lightly-utils==0.0.2
lightning==2.5.2
lightning-utilities==0.15.1
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.6.3
mypy==1.17.1
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbmake==1.5.5
nbsphinx==0.9.7
nest-asyncio==1.6.0
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pycocotools==2.0.10
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pyproj==3.7.1
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytorch-sphinx-theme==0.0.19
pytz==2025.2
pyvista==0.45.3
pywin32==311
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq==27.0.1
rasterio==1.4.3
referencing==0.36.2
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
rich==14.1.0
rpds-py==0.26.0
rtree==1.4.0
ruff==0.12.7
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
scooby==0.10.1
segmentation_models_pytorch==0.5.0
shapely==2.1.1
six==1.17.0
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==5.3.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
sympy @ file:///C:/b/abs_b4u17p23yg/croot/sympy_1738108511395/work
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.19
tinycss2==1.4.0
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchgeo==0.6.2
torchmetrics==1.8.0
torchvision==0.20.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typeshed_client==2.8.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
vtk==9.4.2
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.14
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
yarl==1.20.1
